Complete broker reliability refactor
