+++ scripts/ops/macos-bootstrap.mjs
@@ -458,6 +458,7 @@ function buildAdminEnv(paths, options, seedBrokerEnv) {
     NODE_ENV: "production",
     PATH: "/opt/homebrew/opt/node@24/bin:/opt/homebrew/bin:/usr/bin:/bin",
     PORT: "3000",
+    ADMIN_BASE_URL: "http://127.0.0.1:3000",
     WORKER_PORT: "3001",
     WORKER_BIND_HOST: "127.0.0.1",
     WORKER_BASE_URL: "http://127.0.0.1:3001",
@@ -477,6 +478,7 @@ function buildAdminEnv(paths, options, seedBrokerEnv) {
     CODEX_APP_SERVER_PORT: "4590",
     ADMIN_LAUNCHD_LABEL: options.adminLabel,
     WORKER_LAUNCHD_LABEL: options.workerLabel,
+    ADMIN_PLIST_PATH: paths.adminPlistPath,
     RELEASE_REPO_ROOT: paths.repoRoot,
     RELEASES_ROOT: paths.releasesRoot,
     CURRENT_RELEASE_PATH: paths.currentReleasePath,
@@ -492,6 +494,7 @@ function buildWorkerEnv(paths, options, seedBrokerEnv) {
     NODE_ENV: "production",
     PATH: "/opt/homebrew/opt/node@24/bin:/opt/homebrew/bin:/usr/bin:/bin",
     PORT: "3001",
+    ADMIN_BASE_URL: "http://127.0.0.1:3000",
     WORKER_PORT: "3001",
     WORKER_BIND_HOST: "127.0.0.1",
     WORKER_BASE_URL: "http://127.0.0.1:3001",
@@ -511,6 +514,8 @@ function buildWorkerEnv(paths, options, seedBrokerEnv) {
     CODEX_APP_SERVER_PORT: "4590",
     ADMIN_LAUNCHD_LABEL: options.adminLabel,
     WORKER_LAUNCHD_LABEL: options.workerLabel,
+    ADMIN_PLIST_PATH: paths.adminPlistPath,
+    WORKER_PLIST_PATH: paths.workerPlistPath,
     RELEASE_REPO_ROOT: paths.repoRoot,
     RELEASES_ROOT: paths.releasesRoot,
     CURRENT_RELEASE_PATH: paths.currentReleasePath,
